Refactor pure date helpers into util folder
